--- nxc/protocols/http/__init__.py
@@ -1,3 +1,0 @@
-from .http import http
-
-protocol_object = http
